+++ internal/db/db.go
@@ -36,5 +36,5 @@ func init() {
 		os.Exit(1)
 	}
 
-	err = Conn.AutoMigrate(&UserBio{}, &ReactionRole{}, &userBan{}, &userKick{}, &VerificationFail{}, &UserMute{})
+	err = Conn.AutoMigrate(&UserBio{}, &ReactionRole{}, &userBan{}, &userKick{}, &VerificationFail{}, &UserMute{}, &AnalyticsEvent{})
 }